Split editor and viewer source files
