+++ novelwriter/gui_doc/editor.py
@@ -26,18 +26,14 @@ import logging
 
 from enum import Enum, IntFlag
 from time import time
-from typing import NamedTuple
 
 from PyQt6 import sip
 from PyQt6.QtCore import (
     QAbstractAnimation,
     QMimeData,
-    QObject,
     QPoint,
     QPropertyAnimation,
     QRect,
-    QRegularExpression,
-    QRunnable,
     Qt,
     QTimer,
     QVariant,
@@ -45,7 +41,6 @@ from PyQt6.QtCore import (
     pyqtSlot,
 )
 from PyQt6.QtGui import (
-    QAction,
     QCursor,
     QDragEnterEvent,
     QDragMoveEvent,
@@ -55,7 +50,6 @@ from PyQt6.QtGui import (
     QKeySequence,
     QMouseEvent,
     QPalette,
-    QPixmap,
     QResizeEvent,
     QShortcut,
     QTextBlock,
@@ -65,23 +59,11 @@ from PyQt6.QtGui import (
     QTextFormat,
     QTextOption,
 )
-from PyQt6.QtWidgets import (
-    QApplication,
-    QFrame,
-    QGridLayout,
-    QHBoxLayout,
-    QLabel,
-    QLineEdit,
-    QMenu,
-    QTextEdit,
-    QToolBar,
-    QVBoxLayout,
-    QWidget,
-)
+from PyQt6.QtWidgets import QApplication, QFrame, QMenu, QTextEdit, QWidget
 
 from novelwriter import CONFIG, SHARED
 from novelwriter.common import decodeMimeHandles, fontMatcher, minmax, qtAddAction, qtAddMenu, qtLambda, transferCase
-from novelwriter.constants import nwConst, nwKeyWords, nwLabels, nwShortcode, nwStats, nwStyles, nwUnicode, trStats
+from novelwriter.constants import nwConst, nwKeyWords, nwShortcode, nwStyles, nwUnicode
 from novelwriter.core.document import NWDocument
 from novelwriter.dialogs.editlabel import GuiEditLabel
 from novelwriter.enum import (
@@ -92,39 +74,34 @@ from novelwriter.enum import (
     nwDocMode,
     nwItemClass,
     nwItemType,
-    nwState,
     nwVimMode,
 )
-from novelwriter.extensions.configlayout import NPathColorLabel
 from novelwriter.extensions.eventfilters import WheelEventFilter
-from novelwriter.extensions.modified import NIconToggleButton, NIconToolButton
+from novelwriter.gui_doc.autoreplace import TextAutoReplace
+from novelwriter.gui_doc.completer import CommandCompleter
+from novelwriter.gui_doc.editfooter import GuiDocEditFooter
+from novelwriter.gui_doc.editheader import GuiDocEditHeader
 from novelwriter.gui_doc.editordocument import GuiTextDocument
+from novelwriter.gui_doc.editsearch import GuiDocEditSearch
+from novelwriter.gui_doc.edittoolbar import GuiDocToolBar
 from novelwriter.gui_doc.highlighter import BLOCK_META, BLOCK_TITLE
-from novelwriter.gui_doc.textblock import TextBlockData, spellCheckText
-from novelwriter.gui_side.theme import STYLES_MIN_TOOLBUTTON
+from novelwriter.gui_doc.runnables import BackgroundSpellCheck, BackgroundWordCounter
+from novelwriter.gui_doc.textblock import TextBlockData
 from novelwriter.text.counting import standardCounter
 from novelwriter.text.formats import processHeading
 from novelwriter.tools.lipsum import GuiLipsum
 from novelwriter.types import (
-    QtAlignCenterTop,
     QtAlignJustify,
-    QtAlignLeft,
-    QtAlignLeftTop,
-    QtAlignMiddle,
-    QtAlignRight,
-    QtBlack,
     QtImCurrentSelection,
     QtImCursorRectangle,
     QtKeepAnchor,
     QtKeyDown,
     QtKeyEnter,
-    QtKeyEscape,
     QtKeyLeft,
     QtKeyPageDown,
     QtKeyPageUp,
     QtKeyReturn,
     QtKeyRight,
-    QtKeyTab,
     QtKeyUp,
     QtModCtrl,
     QtModNone,
@@ -2845,1249 +2822,6 @@ class GuiDocEditor(QTextEdit):
                     cursor.setPosition(block.position())
                     self.setTextCursor(cursor)
                     break
-
-
-class CompleterAction(NamedTuple):
-    """Values needed to complete a completer action."""
-
-    pos: int
-    length: int
-    value: str
-
-
-class CommandCompleter(QMenu):
-    """GuiWidget: Command Completer Menu.
-
-    This is a context menu with options populated from the user's
-    defined tags and keys. It also helps to type the meta data keyword
-    on a new line starting with @ or %. The update functions should be
-    called on every keystroke on a line starting with @ or %.
-    """
-
-    __slots__ = ("_parent",)
-
-    insertText = pyqtSignal(int, int, str)
-
-    def __init__(self, parent: QWidget) -> None:
-        super().__init__(parent=parent)
-        self._parent = parent
-        self.triggered.connect(self._emitComplete)
-
-    def updateMetaText(self, text: str, pos: int) -> bool:
-        """Update the menu options based on the line of meta text."""
-        self.clear()
-        kw, sep, _ = text.partition(":")
-        if pos <= len(kw):
-            offset = 0
-            length = len(kw.rstrip())
-            suffix = "" if sep else ":"
-            options = sorted(filter(lambda x: x.startswith(kw.rstrip()), nwKeyWords.VALID_KEYS))
-        else:
-            status, tBits, tPos = SHARED.project.index.scanThis(text)
-            if not status:
-                return False
-            index = bisect.bisect_right(tPos, pos) - 1
-            lookup = tBits[index].lower() if index > 0 else ""
-            offset = tPos[index] if lookup else pos
-            length = len(lookup)
-            suffix = ""
-            options = sorted(filter(lambda x: lookup in x.lower(), SHARED.project.index.getKeyWordTags(kw.strip())))[
-                :15
-            ]
-
-        if not options:
-            return False
-
-        for value in options:
-            rep = value + suffix
-            action = qtAddAction(self, value)
-            action.setData(CompleterAction(pos=offset, length=length, value=rep))
-
-        return True
-
-    def updateCommentText(self, text: str, pos: int) -> bool:
-        """Update the menu options based on the line of comment text."""
-        self.clear()
-        cmd, sep, _ = text.partition(":")
-        if pos <= len(cmd):
-            clean = text[1:].lstrip()[:6].lower()
-            if clean[:6] == "story.":
-                pre, _, key = cmd.partition(".")
-                offset = len(pre) + 1
-                length = len(key)
-                suffix = "" if sep else ": "
-                options = sorted(
-                    filter(
-                        lambda x: x.startswith(key.rstrip()),
-                        SHARED.project.index.getStoryKeys(),
-                    )
-                )
-            elif clean[:5] == "note.":
-                pre, _, key = cmd.partition(".")
-                offset = len(pre) + 1
-                length = len(key)
-                suffix = "" if sep else ": "
-                options = sorted(
-                    filter(
-                        lambda x: x.startswith(key.rstrip()),
-                        SHARED.project.index.getNoteKeys(),
-                    )
-                )
-            elif pos < 12:
-                offset = 0
-                length = len(cmd.rstrip())
-                suffix = ""
-                options = list(
-                    filter(
-                        lambda x: x.startswith(cmd.rstrip()),
-                        ["%Synopsis: ", "%Short: ", "%Story", "%Note"],
-                    )
-                )
-            else:
-                return False
-
-            if options:
-                for value in options:
-                    rep = value + suffix
-                    action = qtAddAction(self, rep.rstrip(":. "))
-                    action.setData(CompleterAction(pos=offset, length=length, value=rep))
-                return True
-
-        return False
-
-    ##
-    #  Events
-    ##
-
-    def keyPressEvent(self, event: QKeyEvent) -> None:
-        """Capture keypresses and forward most of them to the editor."""
-        key = event.key()
-        if key in (QtKeyUp, QtKeyDown):
-            # Let the menu handle navigation
-            super().keyPressEvent(event)
-        elif key in (QtKeyRight, QtKeyReturn, QtKeyEnter, QtKeyTab):
-            # Activate the selection if there is one, otherwise close the completer
-            if action := self.activeAction():
-                action.trigger()
-            else:
-                self.clear()
-                self.close()
-        elif key in (QtKeyLeft, QtKeyEscape):
-            # Cancel the completer
-            self.clear()
-            self.close()
-        else:
-            # Any other keys, send back to the editor
-            # Also close to release the event lock before forwarding key press (#2510)
-            self.clear()
-            self.close()
-            self._parent.keyPressEvent(event)
-
-    ##
-    #  Internal Slots
-    ##
-
-    @pyqtSlot(QAction)
-    def _emitComplete(self, action: QAction) -> None:
-        """Emit the signal to indicate a selection has been made."""
-        if isinstance(data := action.data(), CompleterAction):
-            self.insertText.emit(data.pos, data.length, data.value)
-
-
-class BackgroundWordCounter(QRunnable):
-    """The Off-GUI Thread Word Counter.
-
-    A runnable for the word counter to be run in the thread pool off the
-    main GUI thread.
-    """
-
-    def __init__(self, docEditor: GuiDocEditor, forSelection: bool = False) -> None:
-        super().__init__()
-        self._docEditor = docEditor
-        self._forSelection = forSelection
-        self._isRunning = False
-        self.signals = BackgroundWordCounterSignals()
-
-    def isRunning(self) -> bool:
-        """Return True if the word counter is already running."""
-        return self._isRunning
-
-    @pyqtSlot()
-    def run(self) -> None:
-        """Overloaded run function for the word counter, forwarding the
-        call to the function that does the actual counting.
-        """
-        self._isRunning = True
-        if self._forSelection:
-            text = self._docEditor.getSelectedText()
-        else:
-            text = self._docEditor.getText()
-
-        cC, wC, pC = standardCounter(text)
-        self.signals.countsReady.emit(cC, wC, pC)
-        self._isRunning = False
-
-
-class BackgroundWordCounterSignals(QObject):
-    """The QRunnable cannot emit a signal, so we need a simple QObject
-    to hold the word counter signal.
-    """
-
-    countsReady = pyqtSignal(int, int, int)
-
-
-class BackgroundSpellCheck(QRunnable):
-    """The Off-GUI Thread Spell Checker.
-
-    A runnable that spell checks a batch of text block snapshots in the
-    thread pool off the main GUI thread. It only receives plain text
-    snapshots, and never touches the text document itself.
-    """
-
-    def __init__(self, jobId: int, payload: list[tuple[int, str, int, list[int] | None]]) -> None:
-        super().__init__()
-        self._jobId = jobId
-        self._payload = payload
-        self.signals = BackgroundSpellCheckSignals()
-
-    @pyqtSlot()
-    def run(self) -> None:
-        """Spell check the text snapshots and emit the results."""
-        self.signals.resultsReady.emit(
-            self._jobId,
-            [(index, spellCheckText(text, offset, utf16Map)) for index, text, offset, utf16Map in self._payload],
-        )
-
-
-class BackgroundSpellCheckSignals(QObject):
-    """The QRunnable cannot emit a signal, so we need a simple QObject
-    to hold the spell check result signal.
-    """
-
-    resultsReady = pyqtSignal(int, object)
-
-
-class TextAutoReplace:
-    """Encapsulates the editor auto replace feature."""
-
-    __slots__ = (
-        "_doPadAfter",
-        "_doPadBefore",
-        "_padAfter",
-        "_padBefore",
-        "_padChar",
-        "_quoteDC",
-        "_quoteDO",
-        "_quoteSC",
-        "_quoteSO",
-        "_replaceDQuote",
-        "_replaceDash",
-        "_replaceDots",
-        "_replaceSQuote",
-    )
-
-    def __init__(self) -> None:
-        self.initSettings()
-
-    def initSettings(self) -> None:
-        """Initialise the auto-replace settings from config."""
-        self._quoteSO = CONFIG.fmtSQuoteOpen
-        self._quoteSC = CONFIG.fmtSQuoteClose
-        self._quoteDO = CONFIG.fmtDQuoteOpen
-        self._quoteDC = CONFIG.fmtDQuoteClose
-
-        self._replaceSQuote = CONFIG.doReplaceSQuote
-        self._replaceDQuote = CONFIG.doReplaceDQuote
-        self._replaceDash = CONFIG.doReplaceDash
-        self._replaceDots = CONFIG.doReplaceDots
-
-        self._padChar = nwUnicode.U_THNBSP if CONFIG.fmtPadThin else nwUnicode.U_NBSP
-        self._padBefore = CONFIG.fmtPadBefore
-        self._padAfter = CONFIG.fmtPadAfter
-        self._doPadBefore = bool(CONFIG.fmtPadBefore)
-        self._doPadAfter = bool(CONFIG.fmtPadAfter)
-
-    def process(self, text: str, cursor: QTextCursor) -> bool:
-        """Auto-replace text elements based on main configuration.
-        Returns True if anything was changed.
-        """
-        aPos = cursor.position()
-        bPos = cursor.positionInBlock()
-        block = cursor.block()
-        length = block.length() - 1
-        if length < 1 or bPos - 1 > length:
-            return False
-
-        cursor.movePosition(QtMoveLeft, QtKeepAnchor, min(4, bPos))
-        last = cursor.selectedText()
-        delete, insert = self._determine(last, bPos)
-
-        check = insert
-        if (
-            self._doPadBefore
-            and check
-            and check in self._padBefore
-            and not (check == ":" and length > 1 and text[0] == "@")
-        ):
-            delete = max(delete, 1)
-            chkPos = len(last) - delete - 1
-            if chkPos >= 0 and last[chkPos].isspace():
-                # Strip existing space before inserting a new (#1061)
-                delete += 1
-            insert = self._padChar + insert
-
-        if (
-            self._doPadAfter
-            and check
-            and check in self._padAfter
-            and not (check == ":" and length > 1 and text[0] == "@")
-        ):
-            delete = max(delete, 1)
-            insert = insert + self._padChar
-
-        if delete > 0:
-            cursor.setPosition(aPos)
-            cursor.movePosition(QtMoveLeft, QtKeepAnchor, delete)
-            cursor.insertText(insert)
-            return True
-
-        return False
-
-    def _determine(self, text: str, pos: int) -> tuple[int, str]:
-        """Determine what to replace, if anything."""
-        t1 = text[-1:]
-        t2 = text[-2:]
-        t3 = text[-3:]
-        t4 = text[-4:]
-
-        if self._replaceDQuote and t1 == '"':
-            # Process Double Quote
-            if (
-                pos == 1
-                or (t2[:1].isspace() and t2.endswith('"'))
-                or (pos == 2 and t2 == '>"')
-                or (pos == 3 and t3 == '>>"')
-                or (pos == 2 and t2 == '_"')
-                or (t3[:1].isspace() and t3.endswith('_"'))
-                or (pos == 3 and t3 in ('**"', '=="', '~~"'))
-                or (t4[:1].isspace() and t4.endswith(('**"', '=="', '~~"')))
-            ):
-                return 1, self._quoteDO
-            else:
-                return 1, self._quoteDC
-
-        if self._replaceSQuote and t1 == "'":
-            # Process Single Quote
-            if (
-                pos == 1
-                or (t2[:1].isspace() and t2.endswith("'"))
-                or (pos == 2 and t2 == ">'")
-                or (pos == 3 and t3 == ">>'")
-                or (pos == 2 and t2 == "_'")
-                or (t3[:1].isspace() and t3.endswith("_'"))
-                or (pos == 3 and t3 in ("**'", "=='", "~~'"))
-                or (t4[:1].isspace() and t4.endswith(("**'", "=='", "~~'")))
-            ):
-                return 1, self._quoteSO
-            else:
-                return 1, self._quoteSC
-
-        if self._replaceDash and t1 == "-":
-            # Process Dashes
-            if t4 == "----":
-                return 4, "\u2015"  # Horizontal bar
-            elif t3 == "---":
-                return 3, "\u2014"  # Long dash
-            elif t2 == "--":
-                return 2, "\u2013"  # Short dash
-            elif t2 == "\u2013-":
-                return 2, "\u2014"  # Long dash
-            elif t2 == "\u2014-":
-                return 2, "\u2015"  # Horizontal bar
-
-        if self._replaceDots and t3 == "...":
-            # Process Dots
-            return 3, "\u2026"  # Ellipsis
-
-        if t1 == "\u2028":  # Line separator
-            # This resolves issue #1150
-            return 1, "\u2029"  # Paragraph separator
-
-        return 0, t1
-
-
-class GuiDocToolBar(QWidget):
-    """The Formatting and Options Fold Out Menu.
-
-    Only used by DocEditor, and is opened by the first button in the
-    header.
-    """
-
-    requestDocAction = pyqtSignal(nwDocAction)
-
-    def __init__(self, docEditor: GuiDocEditor) -> None:
-        super().__init__(parent=docEditor)
-
-        logger.debug("Create: GuiDocToolBar")
-
-        iSz = SHARED.theme.baseIconSize
-        self.setContentsMargins(0, 0, 0, 0)
-
-        # General Buttons
-        # ===============
-
-        self.tbBoldMD = NIconToolButton(self, iSz)
-        self.tbBoldMD.setToolTip(self.tr("Markdown Bold"))
-        self.tbBoldMD.clicked.connect(qtLambda(self.requestDocAction.emit, nwDocAction.MD_BOLD))
-
-        self.tbItalicMD = NIconToolButton(self, iSz)
-        self.tbItalicMD.setToolTip(self.tr("Markdown Italic"))
-        self.tbItalicMD.clicked.connect(qtLambda(self.requestDocAction.emit, nwDocAction.MD_ITALIC))
-
-        self.tbStrikeMD = NIconToolButton(self, iSz)
-        self.tbStrikeMD.setToolTip(self.tr("Markdown Strikethrough"))
-        self.tbStrikeMD.clicked.connect(qtLambda(self.requestDocAction.emit, nwDocAction.MD_STRIKE))
-
-        self.tbMarkMD = NIconToolButton(self, iSz)
-        self.tbMarkMD.setToolTip(self.tr("Markdown Highlight"))
-        self.tbMarkMD.clicked.connect(qtLambda(self.requestDocAction.emit, nwDocAction.MD_MARK))
-
-        self.tbBold = NIconToolButton(self, iSz)
-        self.tbBold.setToolTip(self.tr("Shortcode Bold"))
-        self.tbBold.clicked.connect(qtLambda(self.requestDocAction.emit, nwDocAction.SC_BOLD))
-
-        self.tbItalic = NIconToolButton(self, iSz)
-        self.tbItalic.setToolTip(self.tr("Shortcode Italic"))
-        self.tbItalic.clicked.connect(qtLambda(self.requestDocAction.emit, nwDocAction.SC_ITALIC))
-
-        self.tbStrike = NIconToolButton(self, iSz)
-        self.tbStrike.setToolTip(self.tr("Shortcode Strikethrough"))
-        self.tbStrike.clicked.connect(qtLambda(self.requestDocAction.emit, nwDocAction.SC_STRIKE))
-
-        self.tbUnderline = NIconToolButton(self, iSz)
-        self.tbUnderline.setToolTip(self.tr("Shortcode Underline"))
-        self.tbUnderline.clicked.connect(qtLambda(self.requestDocAction.emit, nwDocAction.SC_ULINE))
-
-        self.tbMark = NIconToolButton(self, iSz)
-        self.tbMark.setToolTip(self.tr("Shortcode Highlight"))
-        self.tbMark.clicked.connect(qtLambda(self.requestDocAction.emit, nwDocAction.SC_MARK))
-
-        self.tbSuperscript = NIconToolButton(self, iSz)
-        self.tbSuperscript.setToolTip(self.tr("Shortcode Superscript"))
-        self.tbSuperscript.clicked.connect(qtLambda(self.requestDocAction.emit, nwDocAction.SC_SUP))
-
-        self.tbSubscript = NIconToolButton(self, iSz)
-        self.tbSubscript.setToolTip(self.tr("Shortcode Subscript"))
-        self.tbSubscript.clicked.connect(qtLambda(self.requestDocAction.emit, nwDocAction.SC_SUB))
-
-        # Assemble
-        # ========
-
-        self.outerBox = QVBoxLayout()
-        self.outerBox.addWidget(self.tbBoldMD)
-        self.outerBox.addWidget(self.tbItalicMD)
-        self.outerBox.addWidget(self.tbStrikeMD)
-        self.outerBox.addWidget(self.tbMarkMD)
-        self.outerBox.addSpacing(4)
-        self.outerBox.addWidget(self.tbBold)
-        self.outerBox.addWidget(self.tbItalic)
-        self.outerBox.addWidget(self.tbStrike)
-        self.outerBox.addWidget(self.tbUnderline)
-        self.outerBox.addWidget(self.tbMark)
-        self.outerBox.addWidget(self.tbSuperscript)
-        self.outerBox.addWidget(self.tbSubscript)
-        self.outerBox.setContentsMargins(4, 4, 4, 4)
-        self.outerBox.setSpacing(4)
-
-        self.setLayout(self.outerBox)
-        self.updateTheme()
-
-        # Starts as Invisible
-        self.setVisible(False)
-
-        logger.debug("Ready: GuiDocToolBar")
-
-    def updateTheme(self) -> None:
-        """Initialise GUI elements that depend on specific settings."""
-        logger.debug("Theme Update: GuiDocToolBar")
-
-        syntax = SHARED.theme.syntaxTheme
-        palette = self.palette()
-        palette.setColor(QPalette.ColorRole.Window, syntax.back)
-        palette.setColor(QPalette.ColorRole.WindowText, syntax.text)
-        palette.setColor(QPalette.ColorRole.Text, syntax.text)
-        self.setPalette(palette)
-
-        self.tbBoldMD.setThemeIcon("fmt_bold", "markdown")
-        self.tbItalicMD.setThemeIcon("fmt_italic", "markdown")
-        self.tbStrikeMD.setThemeIcon("fmt_strike", "markdown")
-        self.tbMarkMD.setThemeIcon("fmt_mark", "markdown")
-        self.tbBold.setThemeIcon("fmt_bold", "shortcode")
-        self.tbItalic.setThemeIcon("fmt_italic", "shortcode")
-        self.tbStrike.setThemeIcon("fmt_strike", "shortcode")
-        self.tbUnderline.setThemeIcon("fmt_underline", "shortcode")
-        self.tbMark.setThemeIcon("fmt_mark", "shortcode")
-        self.tbSuperscript.setThemeIcon("fmt_superscript", "shortcode")
-        self.tbSubscript.setThemeIcon("fmt_subscript", "shortcode")
-
-
-class GuiDocEditSearch(QFrame):
-    """The Embedded Document Search/Replace Feature.
-
-    Only used by DocEditor, and is at a fixed position in the
-    QTextEdit's viewport.
-    """
-
-    def __init__(self, docEditor: GuiDocEditor) -> None:
-        super().__init__(parent=docEditor)
-
-        logger.debug("Create: GuiDocEditSearch")
-
-        self.docEditor = docEditor
-
-        iSz = SHARED.theme.baseIconSize
-
-        self.setContentsMargins(0, 0, 0, 0)
-        self.setAutoFillBackground(True)
-        self.setFrameStyle(QFrame.Shape.StyledPanel | QFrame.Shadow.Plain)
-
-        self.mainBox = QGridLayout(self)
-        self.setLayout(self.mainBox)
-
-        # Text Boxes
-        # ==========
-
-        self.searchBox = QLineEdit(self)
-        self.searchBox.setPlaceholderText(self.tr("Search for"))
-        self.searchBox.returnPressed.connect(self._doSearch)
-
-        self.replaceBox = QLineEdit(self)
-        self.replaceBox.setPlaceholderText(self.tr("Replace with"))
-        self.replaceBox.returnPressed.connect(self._doReplace)
-
-        self.searchOpt = QToolBar(self)
-        self.searchOpt.setToolButtonStyle(Qt.ToolButtonStyle.ToolButtonIconOnly)
-        self.searchOpt.setIconSize(iSz)
-        self.searchOpt.setContentsMargins(0, 0, 0, 0)
-
-        self.searchLabel = QLabel(self.tr("Search"), self)
-        self.searchLabel.setIndent(6)
-
-        self.resultLabel = QLabel("?/?", self)
-
-        self.toggleCase = QAction(self.tr("Case Sensitive"), self)
-        self.toggleCase.setCheckable(True)
-        self.toggleCase.setChecked(CONFIG.searchCase)
-        self.toggleCase.toggled.connect(self._doToggleCase)
-        self.searchOpt.addAction(self.toggleCase)
-
-        self.toggleWord = QAction(self.tr("Whole Words Only"), self)
-        self.toggleWord.setCheckable(True)
-        self.toggleWord.setChecked(CONFIG.searchWord)
-        self.toggleWord.toggled.connect(self._doToggleWord)
-        self.searchOpt.addAction(self.toggleWord)
-
-        self.toggleRegEx = QAction(self.tr("RegEx Mode"), self)
-        self.toggleRegEx.setCheckable(True)
-        self.toggleRegEx.setChecked(CONFIG.searchRegEx)
-        self.toggleRegEx.toggled.connect(self._doToggleRegEx)
-        self.searchOpt.addAction(self.toggleRegEx)
-
-        self.toggleLoop = QAction(self.tr("Loop Search"), self)
-        self.toggleLoop.setCheckable(True)
-        self.toggleLoop.setChecked(CONFIG.searchLoop)
-        self.toggleLoop.toggled.connect(self._doToggleLoop)
-        self.searchOpt.addAction(self.toggleLoop)
-
-        self.toggleProject = QAction(self.tr("Search Next File"), self)
-        self.toggleProject.setCheckable(True)
-        self.toggleProject.setChecked(CONFIG.searchNextFile)
-        self.toggleProject.toggled.connect(self._doToggleProject)
-        self.searchOpt.addAction(self.toggleProject)
-
-        self.searchOpt.addSeparator()
-
-        self.toggleMatchCap = QAction(self.tr("Preserve Case"), self)
-        self.toggleMatchCap.setCheckable(True)
-        self.toggleMatchCap.setChecked(CONFIG.searchMatchCap)
-        self.toggleMatchCap.toggled.connect(self._doToggleMatchCap)
-        self.searchOpt.addAction(self.toggleMatchCap)
-
-        self.searchOpt.addSeparator()
-
-        self.cancelSearch = QAction(self.tr("Close Search"), self)
-        self.cancelSearch.triggered.connect(self.closeSearch)
-        self.searchOpt.addAction(self.cancelSearch)
-
-        # Buttons
-        # =======
-
-        self.showReplace = NIconToggleButton(self, iSz)
-        self.showReplace.toggled.connect(self._doToggleReplace)
-
-        self.searchButton = NIconToolButton(self, iSz)
-        self.searchButton.setToolTip(self.tr("Find in current document"))
-        self.searchButton.clicked.connect(self._doSearch)
-
-        self.replaceButton = NIconToolButton(self, iSz)
-        self.replaceButton.setToolTip(self.tr("Find and replace in current document"))
-        self.replaceButton.clicked.connect(self._doReplace)
-
-        self.mainBox.addWidget(self.searchLabel, 0, 0, 1, 2, QtAlignLeft)
-        self.mainBox.addWidget(self.searchOpt, 0, 2, 1, 3, QtAlignRight)
-        self.mainBox.addWidget(self.showReplace, 1, 0, 1, 1)
-        self.mainBox.addWidget(self.searchBox, 1, 1, 1, 2)
-        self.mainBox.addWidget(self.searchButton, 1, 3, 1, 1)
-        self.mainBox.addWidget(self.resultLabel, 1, 4, 1, 1)
-        self.mainBox.addWidget(self.replaceBox, 2, 1, 1, 2)
-        self.mainBox.addWidget(self.replaceButton, 2, 3, 1, 1)
-
-        self.mainBox.setColumnStretch(0, 0)
-        self.mainBox.setColumnStretch(1, 0)
-        self.mainBox.setColumnStretch(2, 1)
-        self.mainBox.setColumnStretch(3, 0)
-        self.mainBox.setColumnStretch(4, 0)
-        self.mainBox.setColumnStretch(5, 0)
-        self.mainBox.setSpacing(2)
-        self.mainBox.setContentsMargins(6, 6, 6, 6)
-
-        self.searchBox.setFixedWidth(200)
-        self.replaceBox.setFixedWidth(200)
-        self.replaceBox.setVisible(False)
-        self.replaceButton.setVisible(False)
-        self.adjustSize()
-
-        self.updateFont()
-        self.updateTheme()
-
-        logger.debug("Ready: GuiDocEditSearch")
-
-    ##
-    #  Properties
-    ##
-
-    @property
-    def searchText(self) -> str:
-        """Return the current search text."""
-        return self.searchBox.text()
-
-    @property
-    def replaceText(self) -> str:
-        """Return the current replace text."""
-        return self.replaceBox.text()
-
-    ##
-    #  Getters
-    ##
-
-    def getSearchObject(self) -> str | QRegularExpression:
-        """Return the current search text either as text or as a regular
-        expression object.
-        """
-        text = self.searchBox.text()
-        if CONFIG.searchRegEx:
-            rxOpt = QRegularExpression.PatternOption.UseUnicodePropertiesOption
-            if not CONFIG.searchCase:
-                rxOpt |= QRegularExpression.PatternOption.CaseInsensitiveOption
-            regEx = QRegularExpression(text, rxOpt)
-            self._alertSearchValid(regEx.isValid())
-            return regEx
-        return text
-
-    ##
-    #  Setters
-    ##
-
-    def setSearchText(self, text: str | None) -> None:
-        """Open the search bar and set the search text to the text
-        provided, if any.
-        """
-        if not self.isVisible():
-            self.setVisible(True)
-        if text is not None:
-            self.searchBox.setText(text)
-        self.searchBox.setFocus()
-        self.searchBox.selectAll()
-        if CONFIG.searchRegEx:
-            self._alertSearchValid(True)
-
-    def setReplaceText(self, text: str) -> None:
-        """Set the replace text."""
-        self.showReplace.setChecked(True)
-        self.replaceBox.setFocus()
-        self.replaceBox.setText(text)
-
-    def setResultCount(self, currRes: int | None, resCount: int | None) -> None:
-        """Set the count values for the current search."""
-        lim = nwConst.MAX_SEARCH_RESULT
-        numCount = f"{lim:n}+" if (resCount or 0) > lim else f"{resCount:n}"
-        sCurrRes = "?" if currRes is None else str(currRes)
-        sResCount = "?" if resCount is None else numCount
-        minWidth = SHARED.theme.getTextWidth(f"{sResCount}//{sResCount}", SHARED.theme.guiFontSmall)
-        self.resultLabel.setText(f"{sCurrRes}/{sResCount}")
-        self.resultLabel.setMinimumWidth(minWidth)
-        self.adjustSize()
-        self.docEditor.updateDocMargins()
-
-    ##
-    #  Methods
-    ##
-
-    def updateFont(self) -> None:
-        """Update the font settings."""
-        self.setFont(SHARED.theme.guiFont)
-        self.searchBox.setFont(SHARED.theme.guiFontSmall)
-        self.replaceBox.setFont(SHARED.theme.guiFontSmall)
-        self.searchLabel.setFont(SHARED.theme.guiFontSmall)
-        self.resultLabel.setFont(SHARED.theme.guiFontSmall)
-        self.resultLabel.setMinimumWidth(SHARED.theme.getTextWidth("?/?", SHARED.theme.guiFontSmall))
-
-    def updateTheme(self) -> None:
-        """Update theme elements."""
-        logger.debug("Theme Update: GuiDocEditSearch")
-
-        palette = QApplication.palette()
-        self.setPalette(palette)
-        self.searchBox.setPalette(palette)
-        self.replaceBox.setPalette(palette)
-
-        # Set icons
-        self.toggleCase.setIcon(SHARED.theme.getIcon("search_case", "tool"))
-        self.toggleWord.setIcon(SHARED.theme.getIcon("search_word", "tool"))
-        self.toggleRegEx.setIcon(SHARED.theme.getIcon("search_regex", "tool"))
-        self.toggleLoop.setIcon(SHARED.theme.getIcon("search_loop", "tool"))
-        self.toggleProject.setIcon(SHARED.theme.getIcon("search_project", "tool"))
-        self.toggleMatchCap.setIcon(SHARED.theme.getIcon("search_preserve", "tool"))
-        self.cancelSearch.setIcon(SHARED.theme.getIcon("search_cancel", "tool"))
-        self.searchButton.setThemeIcon("search", "action")
-        self.replaceButton.setThemeIcon("search_replace", "apply")
-        self.showReplace.setThemeIcon("unfold", "default")
-
-        # Set stylesheets
-        self.searchOpt.setStyleSheet("QToolBar {padding: 0;}")
-        self.showReplace.setStyleSheet("QToolButton {border: none; background: transparent;}")
-
-    def cycleFocus(self) -> bool:
-        """Cycle focus on tab key press. This just alternates focus
-        between the two input boxes, if the replace box is visible.
-        """
-        if self.searchBox.hasFocus():
-            self.replaceBox.setFocus()
-            return True
-        elif self.replaceBox.hasFocus():
-            self.searchBox.setFocus()
-            return True
-        return False
-
-    def anyFocus(self) -> bool:
-        """Return True if any of the input boxes have focus."""
-        return self.hasFocus() or self.isAncestorOf(QApplication.focusWidget())
-
-    ##
-    #  Public Slots
-    ##
-
-    @pyqtSlot()
-    def closeSearch(self) -> None:
-        """Close the search box."""
-        self.showReplace.setChecked(False)
-        self.setVisible(False)
-        self.docEditor.updateDocMargins()
-        self.docEditor.setFocus()
-
-    ##
-    #  Private Slots
-    ##
-
-    @pyqtSlot()
-    def _doSearch(self) -> None:
-        """Call the search action function for the document editor."""
-        self.docEditor.findNext(goBack=(QApplication.keyboardModifiers() == QtModShift))
-
-    @pyqtSlot()
-    def _doReplace(self) -> None:
-        """Call the replace action function for the document editor."""
-        self.docEditor.replaceNext()
-
-    @pyqtSlot(bool)
-    def _doToggleReplace(self, state: bool) -> None:
-        """Toggle the show/hide of the replace box."""
-        self.replaceBox.setVisible(state)
-        self.replaceButton.setVisible(state)
-        self.adjustSize()
-        self.docEditor.updateDocMargins()
-
-    @pyqtSlot(bool)
-    def _doToggleCase(self, state: bool) -> None:
-        """Enable/disable case sensitive mode."""
-        CONFIG.searchCase = state
-
-    @pyqtSlot(bool)
-    def _doToggleWord(self, state: bool) -> None:
-        """Enable/disable whole word search mode."""
-        CONFIG.searchWord = state
-
-    @pyqtSlot(bool)
-    def _doToggleRegEx(self, state: bool) -> None:
-        """Enable/disable regular expression search mode."""
-        CONFIG.searchRegEx = state
-
-    @pyqtSlot(bool)
-    def _doToggleLoop(self, state: bool) -> None:
-        """Enable/disable looping the search."""
-        CONFIG.searchLoop = state
-
-    @pyqtSlot(bool)
-    def _doToggleProject(self, state: bool) -> None:
-        """Enable/disable continuing search in next project file."""
-        CONFIG.searchNextFile = state
-
-    @pyqtSlot(bool)
-    def _doToggleMatchCap(self, state: bool) -> None:
-        """Enable/disable preserving capitalisation when replacing."""
-        CONFIG.searchMatchCap = state
-
-    ##
-    #  Internal Functions
-    ##
-
-    def _alertSearchValid(self, isValid: bool) -> None:
-        """Highlight the search box to indicate the search string is or
-        isn't valid. Take the colour from the replace box.
-        """
-        palette = self.replaceBox.palette()
-        palette.setColor(QPalette.ColorRole.Text, palette.text().color() if isValid else SHARED.theme.errorText)
-        self.searchBox.setPalette(palette)
-
-
-class GuiDocEditHeader(QWidget):
-    """The Embedded Document Header.
-
-    Only used by DocEditor, and is at a fixed position in the
-    QTextEdit's viewport.
-    """
-
-    closeDocumentRequest = pyqtSignal()
-    toggleToolBarRequest = pyqtSignal()
-
-    def __init__(self, docEditor: GuiDocEditor) -> None:
-        super().__init__(parent=docEditor)
-
-        logger.debug("Create: GuiDocEditHeader")
-
-        self.docEditor = docEditor
-
-        self._docHandle = None
-        self._docOutline: dict[int, str] = {}
-        self._state = nwState.NORMAL
-
-        iPx = SHARED.theme.baseIconHeight
-        iSz = SHARED.theme.baseIconSize
-        fPx = SHARED.theme.fontPixelSize
-
-        # Main Widget Settings
-        self.setAutoFillBackground(True)
-
-        # Title Label
-        self.itemTitle = NPathColorLabel("", self)
-        self.itemTitle.setMargin(0)
-        self.itemTitle.setContentsMargins(0, 0, 0, 0)
-        self.itemTitle.setAutoFillBackground(True)
-        self.itemTitle.setAlignment(QtAlignCenterTop)
-        self.itemTitle.setFixedHeight(fPx)
-        self.itemTitle.linkActivated.connect(self._processLabelLink)
-
-        # Other Widgets
-        self.outlineMenu = QMenu(self)
-
-        # Buttons
-        self.tbButton = NIconToolButton(self, iSz)
-        self.tbButton.setVisible(False)
-        self.tbButton.setToolTip(self.tr("Toggle Tool Bar"))
-        self.tbButton.clicked.connect(qtLambda(self.toggleToolBarRequest.emit))
-
-        self.outlineButton = NIconToolButton(self, iSz)
-        self.outlineButton.setVisible(False)
-        self.outlineButton.setToolTip(self.tr("Outline"))
-        self.outlineButton.setMenu(self.outlineMenu)
-
-        self.searchButton = NIconToolButton(self, iSz)
-        self.searchButton.setVisible(False)
-        self.searchButton.setToolTip(self.tr("Search"))
-        self.searchButton.clicked.connect(self.docEditor.toggleSearch)
-
-        self.minmaxButton = NIconToolButton(self, iSz)
-        self.minmaxButton.setVisible(False)
-        self.minmaxButton.setToolTip(self.tr("Toggle Focus Mode"))
-        self.minmaxButton.clicked.connect(qtLambda(self.docEditor.toggleFocusModeRequest.emit))
-
-        self.closeButton = NIconToolButton(self, iSz)
-        self.closeButton.setVisible(False)
-        self.closeButton.setToolTip(self.tr("Close"))
-        self.closeButton.clicked.connect(self._closeDocument)
-
-        # Assemble Layout
-        self.outerBox = QHBoxLayout()
-        self.outerBox.addWidget(self.tbButton, 0, QtAlignMiddle)
-        self.outerBox.addWidget(self.outlineButton, 0, QtAlignMiddle)
-        self.outerBox.addWidget(self.searchButton, 0, QtAlignMiddle)
-        self.outerBox.addSpacing(4)
-        self.outerBox.addWidget(self.itemTitle, 1, QtAlignMiddle)
-        self.outerBox.addSpacing(4)
-        self.outerBox.addSpacing(iPx)
-        self.outerBox.addWidget(self.minmaxButton, 0, QtAlignMiddle)
-        self.outerBox.addWidget(self.closeButton, 0, QtAlignMiddle)
-        self.outerBox.setContentsMargins(4, 4, 4, 4)
-        self.outerBox.setSpacing(0)
-
-        self.setLayout(self.outerBox)
-
-        # Other Signals
-        SHARED.focusModeChanged.connect(self._focusModeChanged)
-
-        # Fix Margins and Size
-        # This is needed for high DPI systems. See issue #499.
-        self.setContentsMargins(0, 0, 0, 0)
-        self.setMinimumHeight(fPx + 4)
-
-        self.updateFont()
-        self.updateTheme()
-
-        logger.debug("Ready: GuiDocEditHeader")
-
-    ##
-    #  Methods
-    ##
-
-    def clearHeader(self) -> None:
-        """Clear the header."""
-        self._docHandle = None
-        self._docOutline = {}
-
-        self.itemTitle.setText("")
-        self.outlineMenu.clear()
-        self.tbButton.setVisible(False)
-        self.outlineButton.setVisible(False)
-        self.searchButton.setVisible(False)
-        self.closeButton.setVisible(False)
-        self.minmaxButton.setVisible(False)
-
-    def setOutline(self, data: dict[int, str]) -> None:
-        """Set the document outline dataset."""
-        if data != self._docOutline:
-            tStart = time()
-            self.outlineMenu.clear()
-            for number, text in data.items():
-                action = qtAddAction(self.outlineMenu, text)
-                action.triggered.connect(qtLambda(self._gotoBlock, number))
-            self._docOutline = data
-            logger.debug("Document outline updated in %.3f ms", 1000 * (time() - tStart))
-
-    def updateFont(self) -> None:
-        """Update the font settings."""
-        self.setFont(SHARED.theme.guiFont)
-        self.itemTitle.setFont(SHARED.theme.guiFontSmall)
-
-    def updateTheme(self) -> None:
-        """Update theme elements."""
-        logger.debug("Theme Update: GuiDocEditHeader")
-
-        self.tbButton.setThemeIcon("fmt_toolbar", "action")
-        self.outlineButton.setThemeIcon("list", "action")
-        self.searchButton.setThemeIcon("search", "action")
-        self.minmaxButton.setThemeIcon("maximise", "action")
-        self.closeButton.setThemeIcon("close", "reject")
-
-        buttonStyle = SHARED.theme.getStyleSheet(STYLES_MIN_TOOLBUTTON)
-        self.tbButton.setStyleSheet(buttonStyle)
-        self.outlineButton.setStyleSheet(buttonStyle)
-        self.searchButton.setStyleSheet(buttonStyle)
-        self.minmaxButton.setStyleSheet(buttonStyle)
-        self.closeButton.setStyleSheet(buttonStyle)
-
-        self.matchColors()
-
-    def matchColors(self) -> None:
-        """Update the colours of the widget to match those of the syntax
-        theme rather than the main GUI.
-        """
-        syntax = SHARED.theme.syntaxTheme
-        palette = self.palette()
-        palette.setColor(QPalette.ColorRole.Window, syntax.back)
-        palette.setColor(QPalette.ColorRole.WindowText, syntax.text)
-        palette.setColor(QPalette.ColorRole.Text, syntax.text)
-        self.setPalette(palette)
-        self.itemTitle.setTextColors(
-            color=palette.windowText().color(),
-            faded=SHARED.theme.fadedText,
-            error=SHARED.theme.errorText,
-        )
-
-    def changeFocusState(self, state: bool) -> None:
-        """Toggle focus state."""
-        self._state = nwState.NORMAL if state else nwState.INACTIVE
-        self.itemTitle.setColorState(self._state)
-
-    def flashError(self) -> None:
-        """Flash a red colour for the header for a moment."""
-        self.itemTitle.setColorState(nwState.ERROR)
-        QTimer.singleShot(250, self._resetColourState)
-
-    def setHandle(self, tHandle: str) -> None:
-        """Set the document title from the handle, or alternatively, set
-        the whole document path within the project.
-        """
-        self._docHandle = tHandle
-        if CONFIG.showFullPath:
-            self.itemTitle.setText(SHARED.project.tree.itemPath(tHandle, withName=True))
-        elif item := SHARED.project.tree[tHandle]:
-            self.itemTitle.setText([(item.itemHandle, item.itemName)])
-
-        self.tbButton.setVisible(True)
-        self.searchButton.setVisible(True)
-        self.outlineButton.setVisible(True)
-        self.closeButton.setVisible(True)
-        self.minmaxButton.setVisible(True)
-
-    ##
-    #  Private Slots
-    ##
-
-    @pyqtSlot()
-    def _closeDocument(self) -> None:
-        """Trigger the close editor on the main window."""
-        self.clearHeader()
-        self.closeDocumentRequest.emit()
-
-    @pyqtSlot(int)
-    def _gotoBlock(self, blockNumber: int) -> None:
-        """Move cursor to a specific heading."""
-        self.docEditor.setCursorLine(blockNumber + 1)
-
-    @pyqtSlot(bool)
-    def _focusModeChanged(self, focusMode: bool) -> None:
-        """Update minimise/maximise icon of the Focus Mode button."""
-        self.minmaxButton.setThemeIcon("minimise" if focusMode else "maximise", "action")
-
-    @pyqtSlot()
-    def _resetColourState(self) -> None:
-        """Reset the colour state of the header title."""
-        self.itemTitle.setColorState(self._state)
-
-    @pyqtSlot(str)
-    def _processLabelLink(self, link: str) -> None:
-        """Process an activated link in the label."""
-        if link.startswith("#"):
-            self.docEditor.requestProjectItemSelected.emit(link.lstrip("#"), True)
-
-
-class GuiDocEditFooter(QWidget):
-    """The Embedded Document Footer.
-
-    Only used by DocEditor, and is at a fixed position in the
-    QTextEdit's viewport.
-    """
-
-    def __init__(self, parent: QWidget) -> None:
-        super().__init__(parent=parent)
-
-        logger.debug("Create: GuiDocEditFooter")
-
-        self._tItem = None
-        self._docHandle = None
-
-        iPx = round(0.9 * SHARED.theme.baseIconHeight)
-        fPx = int(0.9 * SHARED.theme.fontPixelSize)
-
-        # Cached Translations
-        self.initSettings()
-        self._trLineCount = self.tr("Line: {0} ({1})")
-        self._trSelectCount = self.tr("Selected: {0}")
-
-        # Main Widget Settings
-        self.setContentsMargins(0, 0, 0, 0)
-        self.setAutoFillBackground(True)
-
-        # Status
-        self.statusIcon = QLabel("", self)
-        self.statusIcon.setFixedHeight(iPx)
-        self.statusIcon.setAlignment(QtAlignLeftTop)
-
-        self.statusText = QLabel("", self)
-        self.statusText.setAutoFillBackground(True)
-        self.statusText.setFixedHeight(fPx)
-        self.statusText.setAlignment(QtAlignLeftTop)
-
-        # Lines
-        self.linesIcon = QLabel("", self)
-        self.linesIcon.setFixedHeight(iPx)
-        self.linesIcon.setAlignment(QtAlignLeftTop)
-
-        self.linesText = QLabel("", self)
-        self.linesText.setAutoFillBackground(True)
-        self.linesText.setFixedHeight(fPx)
-        self.linesText.setAlignment(QtAlignLeftTop)
-
-        # Words
-        self.wordsIcon = QLabel("", self)
-        self.wordsIcon.setFixedHeight(iPx)
-        self.wordsIcon.setAlignment(QtAlignLeftTop)
-
-        self.wordsText = QLabel("", self)
-        self.wordsText.setAutoFillBackground(True)
-        self.wordsText.setFixedHeight(fPx)
-        self.wordsText.setAlignment(QtAlignLeftTop)
-
-        # Vim mode status bar
-        self.vimStatus = QLabel("", self)
-        self.vimStatus.setAutoFillBackground(True)
-        self.vimStatus.setFixedHeight(fPx)
-        self.vimStatus.setAlignment(QtAlignLeftTop)
-        self._vimMode: nwVimMode | None = None
-        self._vimColor = SHARED.theme.getBaseColor("base")
-        self._vimModes = {}
-
-        # Assemble Layout
-        self.outerBox = QHBoxLayout()
-        self.outerBox.setSpacing(4)
-        self.outerBox.addWidget(self.statusIcon)
-        self.outerBox.addWidget(self.statusText)
-        self.outerBox.addStretch(1)
-        self.outerBox.addSpacing(6)
-        self.outerBox.addWidget(self.vimStatus)
-        self.outerBox.addSpacing(6)
-        self.outerBox.addWidget(self.linesIcon)
-        self.outerBox.addWidget(self.linesText)
-        self.outerBox.addSpacing(6)
-        self.outerBox.addWidget(self.wordsIcon)
-        self.outerBox.addWidget(self.wordsText)
-        self.outerBox.setContentsMargins(8, 8, 8, 8)
-
-        self.setLayout(self.outerBox)
-
-        # Fix Margins and Size
-        # This is needed for high DPI systems. See issue #499.
-        self.setContentsMargins(0, 0, 0, 0)
-        self.setMinimumHeight(fPx + 16)
-
-        # Fix the Colours
-        self.updateFont()
-        self.updateTheme()
-
-        # Initialise Info
-        self.updateMainCount(0, False)
-
-        logger.debug("Ready: GuiDocEditFooter")
-
-    ##
-    #  Methods
-    ##
-
-    def initSettings(self) -> None:
-        """Apply user settings."""
-        self._trMainCount = trStats(nwLabels.STATS_DISPLAY[nwStats.CHARS if CONFIG.useCharCount else nwStats.WORDS])
-
-    def updateFont(self) -> None:
-        """Update the font settings."""
-        self.setFont(SHARED.theme.guiFont)
-        self.statusText.setFont(SHARED.theme.guiFontSmall)
-        self.linesText.setFont(SHARED.theme.guiFontSmall)
-        self.wordsText.setFont(SHARED.theme.guiFontSmall)
-        self.vimStatus.setFont(SHARED.theme.guiFontSmall)
-
-    def updateTheme(self) -> None:
-        """Update theme elements."""
-        logger.debug("Theme Update: GuiDocEditFooter")
-
-        iPx = round(0.9 * SHARED.theme.baseIconHeight)
-        self.linesIcon.setPixmap(SHARED.theme.getPixmap("lines", (iPx, iPx)))
-        self.wordsIcon.setPixmap(SHARED.theme.getPixmap("stats", (iPx, iPx)))
-        self.matchColors()
-        self._vimColor = SHARED.theme.getBaseColor("base")
-        self._vimModes = {
-            nwVimMode.NORMAL: (self.tr("NORMAL"), SHARED.theme.getBaseColor("green")),
-            nwVimMode.INSERT: (self.tr("INSERT"), SHARED.theme.getBaseColor("blue")),
-            nwVimMode.VISUAL: (self.tr("VISUAL"), SHARED.theme.getBaseColor("orange")),
-            nwVimMode.V_LINE: (self.tr("V-LINE"), SHARED.theme.getBaseColor("orange")),
-        }
-
-    def matchColors(self) -> None:
-        """Update the colours of the widget to match those of the syntax
-        theme rather than the main GUI.
-        """
-        syntax = SHARED.theme.syntaxTheme
-
-        palette = self.palette()
-        palette.setColor(QPalette.ColorRole.Window, syntax.back)
-        palette.setColor(QPalette.ColorRole.WindowText, syntax.text)
-        palette.setColor(QPalette.ColorRole.Text, syntax.text)
-
-        self.setPalette(palette)
-        self.statusText.setPalette(palette)
-        self.linesText.setPalette(palette)
-        self.wordsText.setPalette(palette)
-
-    def setHandle(self, tHandle: str | None) -> None:
-        """Set the handle that will populate the footer's data."""
-        self._docHandle = tHandle
-        if self._docHandle is None:
-            logger.debug("No handle set, so clearing the editor footer")
-            self._tItem = None
-        else:
-            self._tItem = SHARED.project.tree[self._docHandle]
-
-        self.updateInfo()
-        self.updateMainCount(0, False)
-
-    def updateInfo(self) -> None:
-        """Update the content of text labels."""
-        if self._tItem is None:
-            sIcon = QPixmap()
-            sText = ""
-        else:
-            iPx = round(0.9 * SHARED.theme.baseIconHeight)
-            status, icon = self._tItem.getImportStatus()
-            sIcon = icon.pixmap(iPx, iPx)
-            sText = f"{status} / {self._tItem.describeMe()}"
-
-        self.statusIcon.setPixmap(sIcon)
-        self.statusText.setText(sText)
-
-    def updateLineCount(self, cursor: QTextCursor) -> None:
-        """Update the line and document position counter."""
-        if document := cursor.document():  # pragma: no branch
-            cPos = cursor.position() + 1
-            cLine = cursor.blockNumber() + 1
-            cCount = max(document.characterCount(), 1)
-            self.linesText.setText(self._trLineCount.format(f"{cLine:n}", f"{100 * cPos // cCount:d} %"))
-
-    def updateMainCount(self, count: int, selection: bool) -> None:
-        """Update main counter information."""
-        if selection and count:
-            text = self._trSelectCount.format(f"{count:n}")
-        elif self._tItem:
-            count = self._tItem.mainCount
-            diff = count - self._tItem.initCount
-            text = self._trMainCount.format(f"{count:n}", f"{diff:+n}")
-        else:
-            text = self._trMainCount.format("0", "+0")
-        self.wordsText.setText(text)
-
-    def updateVimModeStatusBar(self, mode: nwVimMode | None) -> None:
-        """Update the vim Mode status information."""
-        if mode is None:
-            self.vimStatus.setText("")
-            self.vimStatus.setVisible(False)
-            self._vimMode = None
-        elif mode != self._vimMode:
-            text, color = self._vimModes.get(mode, ("", QtBlack))
-            palette = self.vimStatus.palette()
-            palette.setColor(QPalette.ColorRole.WindowText, self._vimColor)
-            palette.setColor(QPalette.ColorRole.Window, color)
-            self.vimStatus.setText(f"  {text}  ")
-            self.vimStatus.setPalette(palette)
-            self.vimStatus.setVisible(True)
-            self._vimMode = mode
 
 
 class VimState:
