+++ novelwriter/gui_doc/viewer.py
@@ -37,22 +37,20 @@ from PyQt6.QtGui import (
     QResizeEvent,
     QTextCursor,
 )
-from PyQt6.QtWidgets import QApplication, QFrame, QHBoxLayout, QMenu, QTextBrowser, QToolButton, QWidget
+from PyQt6.QtWidgets import QApplication, QFrame, QMenu, QTextBrowser, QWidget
 
 from novelwriter import CONFIG, SHARED
 from novelwriter.common import decodeMimeHandles, qtAddAction, qtLambda
 from novelwriter.constants import nwConst, nwStyles, nwUnicode
-from novelwriter.enum import nwChange, nwComment, nwDocAction, nwDocMode, nwItemType, nwState
+from novelwriter.enum import nwChange, nwComment, nwDocAction, nwDocMode, nwItemType
 from novelwriter.error import logException
-from novelwriter.extensions.configlayout import NPathColorLabel
 from novelwriter.extensions.eventfilters import WheelEventFilter
-from novelwriter.extensions.modified import NIconToolButton
 from novelwriter.formats.shared import TextDocumentTheme
 from novelwriter.formats.toqdoc import ToQTextDocument
-from novelwriter.gui_side.theme import STYLES_MIN_TOOLBUTTON
+from novelwriter.gui_doc.viewfooter import GuiDocViewFooter
+from novelwriter.gui_doc.viewheader import GuiDocViewHeader
+from novelwriter.gui_doc.viewhistory import GuiDocViewHistory
 from novelwriter.types import (
-    QtAlignCenterTop,
-    QtAlignMiddle,
     QtKeepAnchor,
     QtMoveAnchor,
     QtScrollAlwaysOff,
@@ -483,477 +481,3 @@ class GuiDocViewer(QTextBrowser):
         """Handle text selection at a given location."""
         self.setTextCursor(self.cursorForPosition(pos))
         self._makeSelection(selType)
-
-
-class GuiDocViewHistory:
-    """GUI: Document Viewer History.
-
-    This class holds the navigation history for the viewer panel, which
-    is used for backward/forward navigation.
-    """
-
-    def __init__(self, docViewer: GuiDocViewer) -> None:
-        self.docViewer = docViewer
-        self._navHistory = []
-        self._posHistory = []
-        self._currPos = -1
-        self._prevPos = -1
-
-    def clear(self) -> None:
-        """Clear the view history."""
-        logger.debug("View history cleared")
-        self._navHistory = []
-        self._posHistory = []
-        self._currPos = -1
-        self._prevPos = -1
-
-    def append(self, tHandle: str) -> bool:
-        """Append a document handle and its scroll bar position to the
-        history, but only if the document is different than the current
-        active entry. Any further entries are truncated.
-        """
-        if self._currPos >= 0 and self._currPos < len(self._navHistory) and tHandle == self._navHistory[self._currPos]:
-            logger.debug("Not updating view hsitory")
-            return False
-
-        self._truncateHistory(self._currPos)
-
-        self._navHistory.append(tHandle)
-        self._posHistory.append(0)
-
-        self._prevPos = self._currPos
-        self._currPos = len(self._navHistory) - 1
-        self._updateScrollBar()
-        self._updateNavButtons()
-
-        self._dumpHistory()
-
-        logger.debug("Added '%s' to view history", tHandle)
-
-        return True
-
-    def forward(self) -> None:
-        """Navigate to the next entry in the view history."""
-        newPos = self._currPos + 1
-        if newPos < len(self._navHistory):
-            logger.debug("Move forward in view history")
-            self._prevPos = self._currPos
-            self._updateScrollBar()
-            self.docViewer.loadText(self._navHistory[newPos], updateHistory=False)
-            self.docViewer.setScrollPosition(self._posHistory[newPos])
-            self._currPos = newPos
-            self._updateNavButtons()
-            self._dumpHistory()
-
-    def backward(self) -> None:
-        """Navigate to the previous entry in the view history."""
-        newPos = self._currPos - 1
-        if newPos >= 0:
-            logger.debug("Move backward in view history")
-            self._prevPos = self._currPos
-            self._updateScrollBar()
-            self.docViewer.loadText(self._navHistory[newPos], updateHistory=False)
-            self.docViewer.setScrollPosition(self._posHistory[newPos])
-            self._currPos = newPos
-            self._updateNavButtons()
-            self._dumpHistory()
-
-    ##
-    #  Internal Functions
-    ##
-
-    def _updateScrollBar(self) -> None:
-        """Update the scrollbar position of the previous entry."""
-        if self._prevPos >= 0 and self._prevPos < len(self._posHistory):
-            self._posHistory[self._prevPos] = self.docViewer.scrollPosition
-
-    def _updateNavButtons(self) -> None:
-        """Update the navigation buttons in the document header."""
-        self.docViewer.docHeader.updateNavButtons(0, len(self._navHistory) - 1, self._currPos)
-
-    def _truncateHistory(self, atPos: int) -> None:
-        """Truncate the navigation history to the given position. Also
-        enforces a maximum length of the navigation history to 20.
-        """
-        nSkip = 1 if atPos > 19 else 0
-        self._navHistory = self._navHistory[nSkip : atPos + 1]
-        self._posHistory = self._posHistory[nSkip : atPos + 1]
-        self._currPos -= nSkip
-        self._prevPos -= nSkip
-
-    def _dumpHistory(self) -> None:
-        """Debug function to dump history to the logger. Since it is a
-        for loop, it is skipped entirely if log level isn't DEBUG.
-        """
-        if CONFIG.isDebug:  # pragma: no cover
-            for i, (h, p) in enumerate(zip(self._navHistory, self._posHistory, strict=False)):
-                a = ">" if i == self._currPos else " "
-                logger.debug(f"History {i + 1:02d}: {a} {h:13s} [x:{p}]")  # noqa: G004
-
-
-class GuiDocViewHeader(QWidget):
-    """The Embedded Document Header.
-
-    Only used by DocViewer, and is at a fixed position in the
-    QTextBrowser's viewport.
-    """
-
-    def __init__(self, docViewer: GuiDocViewer) -> None:
-        super().__init__(parent=docViewer)
-
-        logger.debug("Create: GuiDocViewHeader")
-
-        self.docViewer = docViewer
-
-        # Internal Variables
-        self._docHandle = None
-        self._docOutline: dict[str, tuple[str, int]] = {}
-
-        iSz = SHARED.theme.baseIconSize
-        fPx = SHARED.theme.fontPixelSize
-
-        # Main Widget Settings
-        self.setAutoFillBackground(True)
-
-        # Title Label
-        self.itemTitle = NPathColorLabel("", self, faded=SHARED.theme.fadedText)
-        self.itemTitle.setMargin(0)
-        self.itemTitle.setContentsMargins(0, 0, 0, 0)
-        self.itemTitle.setAutoFillBackground(True)
-        self.itemTitle.setAlignment(QtAlignCenterTop)
-        self.itemTitle.setFixedHeight(fPx)
-        self.itemTitle.linkActivated.connect(self._processLabelLink)
-
-        # Other Widgets
-        self.outlineMenu = QMenu(self)
-
-        # Buttons
-        self.outlineButton = NIconToolButton(self, iSz)
-        self.outlineButton.setVisible(False)
-        self.outlineButton.setToolTip(self.tr("Outline"))
-        self.outlineButton.setMenu(self.outlineMenu)
-
-        self.backButton = NIconToolButton(self, iSz)
-        self.backButton.setVisible(False)
-        self.backButton.setToolTip(self.tr("Go Backward"))
-        self.backButton.clicked.connect(self.docViewer.navBackward)
-
-        self.forwardButton = NIconToolButton(self, iSz)
-        self.forwardButton.setVisible(False)
-        self.forwardButton.setToolTip(self.tr("Go Forward"))
-        self.forwardButton.clicked.connect(self.docViewer.navForward)
-
-        self.editButton = NIconToolButton(self, iSz)
-        self.editButton.setVisible(False)
-        self.editButton.setToolTip(self.tr("Open in Editor"))
-        self.editButton.clicked.connect(self._editDocument)
-
-        self.refreshButton = NIconToolButton(self, iSz)
-        self.refreshButton.setVisible(False)
-        self.refreshButton.setToolTip(self.tr("Reload"))
-        self.refreshButton.clicked.connect(self._refreshDocument)
-
-        self.closeButton = NIconToolButton(self, iSz)
-        self.closeButton.setVisible(False)
-        self.closeButton.setToolTip(self.tr("Close"))
-        self.closeButton.clicked.connect(self._closeDocument)
-
-        # Assemble Layout
-        self.outerBox = QHBoxLayout()
-        self.outerBox.addWidget(self.outlineButton, 0, QtAlignMiddle)
-        self.outerBox.addWidget(self.backButton, 0, QtAlignMiddle)
-        self.outerBox.addWidget(self.forwardButton, 0, QtAlignMiddle)
-        self.outerBox.addSpacing(4)
-        self.outerBox.addWidget(self.itemTitle, 1, QtAlignMiddle)
-        self.outerBox.addSpacing(4)
-        self.outerBox.addWidget(self.editButton, 0, QtAlignMiddle)
-        self.outerBox.addWidget(self.refreshButton, 0, QtAlignMiddle)
-        self.outerBox.addWidget(self.closeButton, 0, QtAlignMiddle)
-        self.outerBox.setSpacing(0)
-
-        self.setLayout(self.outerBox)
-
-        # Fix Margins and Size
-        # This is needed for high DPI systems. See issue #499.
-        self.setContentsMargins(0, 0, 0, 0)
-        self.outerBox.setContentsMargins(4, 4, 4, 4)
-        self.setMinimumHeight(fPx + 4)
-
-        self.updateFont()
-        self.updateTheme()
-
-        logger.debug("Ready: GuiDocViewHeader")
-
-    ##
-    #  Methods
-    ##
-
-    def clearHeader(self) -> None:
-        """Clear the header."""
-        self._docHandle = None
-        self._docOutline = {}
-
-        self.itemTitle.setText("")
-        self.outlineMenu.clear()
-        self.outlineButton.setVisible(False)
-        self.backButton.setVisible(False)
-        self.forwardButton.setVisible(False)
-        self.editButton.setVisible(False)
-        self.refreshButton.setVisible(False)
-        self.closeButton.setVisible(False)
-
-    def setOutline(self, data: dict[str, tuple[str, int]]) -> None:
-        """Set the document outline dataset."""
-        tHandle = self._docHandle
-        if data != self._docOutline and tHandle:
-            self.outlineMenu.clear()
-            entries = []
-            minLevel = 5
-            for title, (text, level) in data.items():
-                if title != "T0000":
-                    entries.append((title, text, level))
-                    minLevel = min(minLevel, level)
-            for title, text, level in entries[:30]:
-                indent = "    " * (level - minLevel)
-                action = qtAddAction(self.outlineMenu, f"{indent}{text}")
-                action.triggered.connect(lambda _, title=title: self.docViewer.navigateTo(f"#{tHandle}:{title}"))
-            self._docOutline = data
-
-    def updateFont(self) -> None:
-        """Update the font settings."""
-        self.setFont(SHARED.theme.guiFont)
-        self.itemTitle.setFont(SHARED.theme.guiFontSmall)
-
-    def updateTheme(self) -> None:
-        """Update theme elements."""
-        logger.debug("Theme Update: GuiDocViewHeader")
-
-        self.outlineButton.setThemeIcon("list", "action")
-        self.backButton.setThemeIcon("chevron_left", "action")
-        self.forwardButton.setThemeIcon("chevron_right", "action")
-        self.editButton.setThemeIcon("edit", "change")
-        self.refreshButton.setThemeIcon("refresh", "change")
-        self.closeButton.setThemeIcon("close", "reject")
-
-        buttonStyle = SHARED.theme.getStyleSheet(STYLES_MIN_TOOLBUTTON)
-        self.outlineButton.setStyleSheet(buttonStyle)
-        self.backButton.setStyleSheet(buttonStyle)
-        self.forwardButton.setStyleSheet(buttonStyle)
-        self.editButton.setStyleSheet(buttonStyle)
-        self.refreshButton.setStyleSheet(buttonStyle)
-        self.closeButton.setStyleSheet(buttonStyle)
-
-        self.matchColors()
-
-    def matchColors(self) -> None:
-        """Update the colours of the widget to match those of the syntax
-        theme rather than the main GUI.
-        """
-        syntax = SHARED.theme.syntaxTheme
-        palette = self.palette()
-        palette.setColor(QPalette.ColorRole.Window, syntax.back)
-        palette.setColor(QPalette.ColorRole.WindowText, syntax.text)
-        palette.setColor(QPalette.ColorRole.Text, syntax.text)
-        self.setPalette(palette)
-        self.itemTitle.setTextColors(
-            color=palette.windowText().color(),
-            faded=SHARED.theme.fadedText,
-        )
-
-    def changeFocusState(self, state: bool) -> None:
-        """Toggle focus state."""
-        self.itemTitle.setColorState(nwState.NORMAL if state else nwState.INACTIVE)
-
-    def setHandle(self, tHandle: str) -> None:
-        """Set the document title from the handle, or alternatively, set
-        the whole document path within the project.
-        """
-        self._docHandle = tHandle
-        if CONFIG.showFullPath:
-            self.itemTitle.setText(SHARED.project.tree.itemPath(tHandle, withName=True))
-        elif item := SHARED.project.tree[tHandle]:
-            self.itemTitle.setText([(item.itemHandle, item.itemName)])
-
-        self.backButton.setVisible(True)
-        self.forwardButton.setVisible(True)
-        self.outlineButton.setVisible(True)
-        self.editButton.setVisible(True)
-        self.refreshButton.setVisible(True)
-        self.closeButton.setVisible(True)
-
-    def updateNavButtons(self, firstIdx: int, lastIdx: int, currIdx: int) -> None:
-        """Enable and disable nav buttons based on index in history."""
-        self.backButton.setEnabled(currIdx > firstIdx)
-        self.forwardButton.setEnabled(currIdx < lastIdx)
-
-    ##
-    #  Private Slots
-    ##
-
-    @pyqtSlot()
-    def _closeDocument(self) -> None:
-        """Trigger the close editor/viewer on the main window."""
-        self.clearHeader()
-        self.docViewer.closeDocumentRequest.emit()
-
-    @pyqtSlot()
-    def _refreshDocument(self) -> None:
-        """Reload the content of the document."""
-        self.docViewer.reloadDocumentRequest.emit()
-
-    @pyqtSlot()
-    def _editDocument(self) -> None:
-        """Open the document in the editor."""
-        if tHandle := self._docHandle:
-            self.docViewer.openDocumentRequest.emit(tHandle, nwDocMode.EDIT, "", True)
-
-    @pyqtSlot(str)
-    def _processLabelLink(self, link: str) -> None:
-        """Process an activated link in the label."""
-        if link.startswith("#"):
-            self.docViewer.requestProjectItemSelected.emit(link.lstrip("#"), True)
-
-
-class GuiDocViewFooter(QWidget):
-    """The Embedded Document Footer.
-
-    Only used by DocViewer, and is at a fixed position in the
-    QTextBrowser's viewport.
-    """
-
-    def __init__(self, docViewer: GuiDocViewer) -> None:
-        super().__init__(parent=docViewer)
-
-        logger.debug("Create: GuiDocViewFooter")
-
-        self.docViewer = docViewer
-
-        # Internal Variables
-        self._docHandle = None
-
-        iPx = SHARED.theme.baseIconHeight
-        iSz = SHARED.theme.baseIconSize
-
-        # Main Widget Settings
-        self.setContentsMargins(0, 0, 0, 0)
-        self.setAutoFillBackground(True)
-
-        # Show/Hide Details
-        self.showHide = NIconToolButton(self, iSz)
-        self.showHide.setToolTip(self.tr("Show/Hide Viewer Panel"))
-        self.showHide.clicked.connect(lambda: self.docViewer.togglePanelVisibility.emit())
-
-        # Show Comments
-        self.showComments = QToolButton(self)
-        self.showComments.setText(self.tr("Comments"))
-        self.showComments.setToolTip(self.tr("Show Comments"))
-        self.showComments.setCheckable(True)
-        self.showComments.setChecked(CONFIG.viewComments)
-        self.showComments.setToolButtonStyle(Qt.ToolButtonStyle.ToolButtonTextBesideIcon)
-        self.showComments.setIconSize(iSz)
-        self.showComments.toggled.connect(self._doToggleComments)
-
-        # Show Synopsis
-        self.showSynopsis = QToolButton(self)
-        self.showSynopsis.setText(self.tr("Synopsis"))
-        self.showSynopsis.setToolTip(self.tr("Show Synopsis Comments"))
-        self.showSynopsis.setCheckable(True)
-        self.showSynopsis.setChecked(CONFIG.viewSynopsis)
-        self.showSynopsis.setToolButtonStyle(Qt.ToolButtonStyle.ToolButtonTextBesideIcon)
-        self.showSynopsis.setIconSize(iSz)
-        self.showSynopsis.toggled.connect(self._doToggleSynopsis)
-
-        # Show Notes
-        self.showNotes = QToolButton(self)
-        self.showNotes.setText(self.tr("Notes"))
-        self.showNotes.setToolTip(self.tr("Show Notes"))
-        self.showNotes.setCheckable(True)
-        self.showNotes.setChecked(CONFIG.viewNotes)
-        self.showNotes.setToolButtonStyle(Qt.ToolButtonStyle.ToolButtonTextBesideIcon)
-        self.showNotes.setIconSize(iSz)
-        self.showNotes.toggled.connect(self._doToggleNotes)
-
-        # Assemble Layout
-        self.outerBox = QHBoxLayout()
-        self.outerBox.addWidget(self.showHide, 0)
-        self.outerBox.addStretch(1)
-        self.outerBox.addWidget(self.showComments, 0)
-        self.outerBox.addWidget(self.showSynopsis, 0)
-        self.outerBox.addWidget(self.showNotes, 0)
-        self.outerBox.setSpacing(4)
-        self.setLayout(self.outerBox)
-
-        # Fix Margins and Size
-        # This is needed for high DPI systems. See issue #499.
-        self.setContentsMargins(0, 0, 0, 0)
-        self.outerBox.setContentsMargins(4, 4, 4, 4)
-        self.setMinimumHeight(iPx + 8)
-
-        self.updateFont()
-        self.updateTheme()
-
-        logger.debug("Ready: GuiDocViewFooter")
-
-    ##
-    #  Methods
-    ##
-
-    def updateFont(self) -> None:
-        """Update the font settings."""
-        self.setFont(SHARED.theme.guiFont)
-        self.showComments.setFont(SHARED.theme.guiFontSmall)
-        self.showSynopsis.setFont(SHARED.theme.guiFontSmall)
-        self.showNotes.setFont(SHARED.theme.guiFontSmall)
-
-    def updateTheme(self) -> None:
-        """Update theme elements."""
-        logger.debug("Theme Update: GuiDocViewFooter")
-
-        fPx = int(0.9 * SHARED.theme.fontPixelSize)
-        bulletIcon = SHARED.theme.getToggleIcon("bullet", (fPx, fPx), "action")
-
-        self.showHide.setThemeIcon("panel", "default")
-        self.showComments.setIcon(bulletIcon)
-        self.showSynopsis.setIcon(bulletIcon)
-        self.showNotes.setIcon(bulletIcon)
-
-        buttonStyle = SHARED.theme.getStyleSheet(STYLES_MIN_TOOLBUTTON)
-        self.showHide.setStyleSheet(buttonStyle)
-        self.showComments.setStyleSheet(buttonStyle)
-        self.showSynopsis.setStyleSheet(buttonStyle)
-        self.showNotes.setStyleSheet(buttonStyle)
-
-        self.matchColors()
-
-    def matchColors(self) -> None:
-        """Update the colours of the widget to match those of the syntax
-        theme rather than the main GUI.
-        """
-        syntax = SHARED.theme.syntaxTheme
-        palette = self.palette()
-        palette.setColor(QPalette.ColorRole.Window, syntax.back)
-        palette.setColor(QPalette.ColorRole.WindowText, syntax.text)
-        palette.setColor(QPalette.ColorRole.Text, syntax.text)
-        self.setPalette(palette)
-
-    ##
-    #  Private Slots
-    ##
-
-    @pyqtSlot(bool)
-    def _doToggleComments(self, state: bool) -> None:
-        """Toggle the view comment button and reload the document."""
-        CONFIG.viewComments = state
-        self.docViewer.reloadText()
-
-    @pyqtSlot(bool)
-    def _doToggleSynopsis(self, state: bool) -> None:
-        """Toggle the view synopsis button and reload the document."""
-        CONFIG.viewSynopsis = state
-        self.docViewer.reloadText()
-
-    @pyqtSlot(bool)
-    def _doToggleNotes(self, state: bool) -> None:
-        """Toggle the view notes button and reload the document."""
-        CONFIG.viewNotes = state
-        self.docViewer.reloadText()
